oee_analyzer/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
gradio>=4.0.0
huggingface-hub>=0.20.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0 